add ComplexSolver.cs
