+++ BlazorRummiSolve.Tests/ScoreComplexSolverTests.cs
@@ -3,7 +3,7 @@ using RummiSolve.Solver;
 
 namespace BlazorRummiSolve.Tests;
 
-public class IncrementalSolverTests
+public class ScoreComplexSolverTests
 {
     [Fact]
     public void SearchSolution_Valid()
@@ -21,20 +21,15 @@ public class IncrementalSolverTests
             new Tile(10, TileColor.Black),
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.True(won);
-        Assert.True(solution.IsValid);
-        Assert.Equal(3, tilesToPlay.Count);
-        Assert.Equal(0, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(30, bestScore);
     }
 
     [Fact]
@@ -53,22 +48,16 @@ public class IncrementalSolverTests
             new Tile(10, TileColor.Red),
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.True(won);
-        Assert.True(solution.IsValid);
-        Assert.Equal(2, tilesToPlay.Count);
-        Assert.Equal(0, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(20, bestScore);
     }
-
 
     [Fact]
     public void SearchSolution_ValidRun()
@@ -84,20 +73,15 @@ public class IncrementalSolverTests
             new Tile(4, TileColor.Red)
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.True(won);
-        Assert.True(solution.IsValid);
-        Assert.Single(tilesToPlay);
-        Assert.Equal(0, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(4, bestScore);
     }
     
     [Fact]
@@ -116,22 +100,17 @@ public class IncrementalSolverTests
             new Tile(13),
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.True(won);
-        Assert.True(solution.IsValid);
-        Assert.Equal(2,tilesToPlay.Count);
-        Assert.Equal(0, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(25, bestScore);
     }
-
+    
     [Fact]
     public void SearchSolution_ValidNotWon()
     {
@@ -155,20 +134,15 @@ public class IncrementalSolverTests
             new Tile(3, TileColor.Red),
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.False(won);
-        Assert.True(solution.IsValid);
-        Assert.Equal(6, tilesToPlay.Count);
-        Assert.Equal(0, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(36, bestScore);
     }
     
     [Fact]
@@ -192,20 +166,15 @@ public class IncrementalSolverTests
             new Tile(13),
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.True(won);
-        Assert.True(solution.IsValid);
-        Assert.Equal(3, tilesToPlay.Count);
-        Assert.Equal(0, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(35, bestScore);
     }
     
     [Fact]
@@ -227,22 +196,17 @@ public class IncrementalSolverTests
             new Tile(true)
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
-        var tilesToPlay = solver.TilesToPlay.ToList();
-        var jokerToPlay = solver.JokerToPlay;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.True(won);
-        Assert.True(solution.IsValid);
-        Assert.Equal(4, tilesToPlay.Count);
-        Assert.Equal(1, jokerToPlay);
+        Assert.True(canPlay);
+        Assert.Equal(35, bestScore);
     }
-
+    
     [Fact]
     public void SearchSolution_Invalid()
     {
@@ -257,15 +221,14 @@ public class IncrementalSolverTests
             new Tile(5)
         ]);
 
-        var solver = IncrementalSolver.Create(boardSet, playerSet);
+        var solver = ScoreComplexSolver.Create(boardSet, playerSet);
 
         // Act
-        solver.SearchSolution();
-        var won = solver.Won;
-        var solution = solver.BestSolution;
+        var canPlay = solver.SearchBestScore();
+        var bestScore = solver.BestScore;
 
         // Assert
-        Assert.False(won);
-        Assert.False(solution.IsValid);
+        Assert.False(canPlay);
+        Assert.Equal(0, bestScore);
     }
 }
