+++ RummiSolve/RummiSolve/Solver/BestScoreBaseSolver.cs
@@ -1,9 +1,12 @@
+using RummiSolve.Solver.Abstract;
 using RummiSolve.Solver.Interfaces;
 
 namespace RummiSolve.Solver;
 
-public class BestScoreFirstSolver : SolverBase, ISolver
+public class BestScoreBaseSolver : BaseSolver, ISolver
 {
+    private readonly bool[] _isPlayerTile;
+    private readonly int _boardJokers;
     private readonly int _availableJokers;
     private int _bestSolutionScore;
 
@@ -13,30 +16,48 @@ public class BestScoreFirstSolver : SolverBase, ISolver
     public int JokerToPlay { get; private set; }
     public bool Won { get; private set; }
 
-    private BestScoreFirstSolver(Tile[] tiles, int jokers) : base(tiles, jokers)
+    private BestScoreBaseSolver(Tile[] tiles, int jokers, bool[] isPlayerTile, int boardJokers) :
+        base(tiles,
+            jokers)
     {
         _availableJokers = jokers;
+        _isPlayerTile = isPlayerTile;
+        _boardJokers = boardJokers;
     }
 
 
-    public static BestScoreFirstSolver Create(Set playerSet)
+    public static BestScoreBaseSolver Create(Set boardSet, Set playerSet)
     {
-        var tiles = new List<Tile>(playerSet.Tiles);
+        var capacity = boardSet.Tiles.Count + playerSet.Tiles.Count;
+        var combined = new List<(Tile tile, bool isPlayerTile)>(capacity);
 
-        tiles.Sort();
+        combined.AddRange(boardSet.Tiles.Select(tile => (tile, false)));
+        combined.AddRange(playerSet.Tiles.Select(tile => (tile, true)));
 
-        if (playerSet.Jokers > 0) tiles.RemoveRange(tiles.Count - playerSet.Jokers, playerSet.Jokers);
-        
-        
-        return new BestScoreFirstSolver(
-            tiles.ToArray(),
-            playerSet.Jokers
+        var totalJokers = boardSet.Jokers + playerSet.Jokers;
+
+        combined.Sort((x, y) =>
+        {
+            var tileCompare = x.tile.CompareTo(y.tile);
+            return tileCompare != 0 ? tileCompare : x.isPlayerTile.CompareTo(y.isPlayerTile);
+        });
+
+        if (totalJokers > 0) combined.RemoveRange(combined.Count - totalJokers, totalJokers);
+
+        var finalTiles = combined.Select(pair => pair.tile).ToArray();
+        var isPlayerTile = combined.Select(pair => pair.isPlayerTile).ToArray();
+
+        return new BestScoreBaseSolver(
+            finalTiles,
+            totalJokers,
+            isPlayerTile,
+            boardSet.Jokers
         );
     }
 
     public void SearchSolution()
     {
-        var scoreSolver = new ScoreFirstSolver(Tiles, Jokers);
+        var scoreSolver = new ScoreComplexSolver(Tiles, Jokers, _isPlayerTile);
 
         var canPlay = scoreSolver.SearchBestScore();
 
@@ -47,14 +68,18 @@ public class BestScoreFirstSolver : SolverBase, ISolver
         _bestSolutionScore = scoreSolver.BestScore;
         BestSolution = FindSolution(new Solution(), 0, 0);
         Won = UsedTiles.All(b => b);
-        TilesToPlay = Tiles.Where((_, i) => UsedTiles[i]);
-        JokerToPlay = _availableJokers - Jokers;
+        TilesToPlay = Tiles.Where((_, i) => _isPlayerTile[i] && UsedTiles[i]);
+        JokerToPlay = _availableJokers - Jokers - _boardJokers;
     }
 
 
     private bool ValidateCondition(int solutionScore)
     {
-        return solutionScore == _bestSolutionScore;
+        if (solutionScore != _bestSolutionScore) return false;
+        var allBoardTilesUsed =
+            !UsedTiles.Where((use, i) => !use && !_isPlayerTile[i]).Any(); //check pas de joker restant ?
+
+        return allBoardTilesUsed;
     }
 
 
@@ -76,7 +101,8 @@ public class BestScoreFirstSolver : SolverBase, ISolver
 
             if (solGroup.IsValid) return solGroup;
 
-            startIndex++;
+            if (_isPlayerTile[startIndex]) startIndex++;
+            else return solution;
         }
 
         return solution;
@@ -87,7 +113,7 @@ public class BestScoreFirstSolver : SolverBase, ISolver
         where TS : ValidSet
     {
         UsedTiles[firstUnusedTileIndex] = true;
-        var firstTileScore = Tiles[firstUnusedTileIndex].Value;
+        var firstTileScore = _isPlayerTile[firstUnusedTileIndex] ? Tiles[firstUnusedTileIndex].Value : 0;
         foreach (var set in sets)
         {
             MarkTilesAsUsedOut(set, firstUnusedTileIndex, out var playerSetScore);
@@ -117,8 +143,6 @@ public class BestScoreFirstSolver : SolverBase, ISolver
         playerSetScore = 0;
         foreach (var tile in set.Tiles.Skip(1))
         {
-            playerSetScore += tile.Value;
-            
             if (tile.IsJoker)
             {
                 Jokers -= 1;
@@ -130,7 +154,12 @@ public class BestScoreFirstSolver : SolverBase, ISolver
                 if (UsedTiles[i] || !Tiles[i].Equals(tile)) continue;
 
                 UsedTiles[i] = true;
-                
+
+                if (_isPlayerTile[i])
+                {
+                    playerSetScore += tile.Value;
+                }
+
                 break;
             }
         }
